Idea to let fileFactories override core modules
