+++ src/PipeDream.js
@@ -5,27 +5,6 @@ export default class PipeDream {
         this.options = options
     }
 
-    /* Idea on how to merge custom modules */
-    /* Usage? */
-    /* Use global prefix ~pd.Attribute */
-    /* Use store */
-    /* Use prefix when importing */
-    modules(overriddenModules) {
-
-        let defaultModules = {            
-            Attribute,
-            AttributeFactory,
-            ObjectModelCollection,
-            ObjectModelEntity,
-            ObjectModelEntityFactory,
-            Segment,
-            SegmentRow,
-            SketchParser
-        }
-
-        return merge(defaultModules, overriddenModules)
-    }
-
     get defaultStore() {
         return {
             state: defaultStore.state(this.options),
